requirements.txt
torch
torchvision
matplotlib
pyyaml
opencv-python
pandas
